[PR-153] Further fixes for Polish EPG encoding
It seems from my debugging that we cannot force the encoding to specific
type as the only one, because the providers are sometimes encoding
the current/next events differently then the rest. As the result the
part of events could have wrong encoding. There are also situations,
where firstly loaded event is OK, but further update is screwing
the encoding on it.

My solution is adding special encoding type: PL_AUTO which cover all
Polish EPG. This way the encoding is correct for both current/next
and further events. Side effect is that the encoding list for Polish
channels is smaller (don't need to check if channel is encoded in
ISO8859-2 or in ISO6937, it just works if it is set to PL_AUTO).

The is still a place to force encoding to specified one for a whole
transponder or specific services, but it is not applicable to Polish
providers.